+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## v2 (2019-08-11)
+
+- Removed draft feature for global blocks like menus and footers
+
 ## 0.1 (2019-08-10)
 
 This is the initial release of the Draft Mode extension for Mobirise.